requirements.txt
absl-py==2.1.0
accelerate==1.6.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrdict==2.0.1
attrs==23.2.0
audioread==3.0.1
av==14.1.0
babel==2.17.0
beautifulsoup4==4.13.3
binpacking==1.5.2
black==24.1.0
bleach==6.2.0
blessed==1.20.0
boto3==1.37.14
botocore==1.37.14
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cog==0.14.6
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cpm-kernels==1.0.11
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
dacite==1.9.2
DataProperty==1.1.0
datasets==2.16.1
debugpy==1.8.13
decorator==4.4.2
decord==0.6.0
deepspeed==0.16.4
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
einx==0.3.0
encodec==0.1.1
environs==14.1.1
et_xmlfile==2.0.0
evaluate==0.4.3
eventlet==0.39.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-gpu-cu12==1.10.0
fastapi==0.115.8
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.17.0
fire==0.7.0
fonttools==4.56.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2023.10.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
gpustat==1.1.1
gradio==5.12.0
gradio_client==1.5.4
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.16.3
httptools==0.6.4
httpx
huggingface-hub==0.29.1
humanfriendly==10.0
humanize==4.12.1
identify==2.6.10
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.13.2
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
latex2sympy2==1.9.1
lazy_loader==0.4
librosa==0.9.0
llava==0.0.1.dev0
llvmlite==0.44.0
loguru==0.7.3
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdurl==0.1.2
mistune==3.1.2
modelscope==1.24.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.15
multiprocessing-logging==0.3.4
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
nncore==0.4.5
nodeenv==1.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numexpr==2.10.2
numpy==1.26.4
openai==1.66.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
oss2==2.19.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
PasteDeploy==3.1.0
pathspec==0.12.1
pathvalidate==3.2.3
peft
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
pre_commit==4.2.0
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==3.20.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.7
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyeclib==1.6.4
Pygments==2.19.1
pyparsing==3.2.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
qwen-omni-utils==0.0.3
qwen-vl-utils==0.0.8
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rouge==1.0.1
rouge-chinese==1.0.3
rpds-py==0.23.1
ruff==0.9.7
s3transfer==0.11.4
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.1.99
sentry-sdk==2.25.1
setproctitle==1.3.5
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
sqlitedict==2.1.0
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.45.3
structlog==24.4.0
swift==2.35.0
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==8.3.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.50.3
transformers-stream-generator==0.0.5
triton==3.2.0
trl
typepy==1.3.4
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vector-quantize-pytorch==1.22.3
virtualenv==20.30.0
vocos==0.1.0
wandb==0.19.9
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.13
xattr==1.1.4
xxhash==3.5.0
yarl==1.18.3
yt-dlp==2025.3.31
zipp==3.21.0
zss==1.2.0
zstandard==0.23.0
flash_attn==2.7.4.post1
